% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromaList.R
\name{chromaList}
\alias{chromaList}
\title{chromaList}
\usage{
chromaList(x, n)
}
\arguments{
\item{x}{a object of featureChroma}

\item{n}{ID of feature}
}
\description{
chromaList
}
\examples{
\dontrun{
chrs_raw <- chromaList()
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
