% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromaList.R
\name{chromaProfile}
\alias{chromaProfile}
\title{chromaProfile}
\usage{
chromaProfile(xy)
}
\arguments{
\item{xy}{a object of featureChroma}
}
\description{
chromaProfile for merictable function
}
\examples{
\dontrun{
chrs_raw <- chromaProfile()
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
