% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatogramReport.R
\name{chromatogramReport}
\alias{chromatogramReport}
\title{chromatogramReport}
\usage{
chromatogramReport(subdir, object, peakInter = NULL, expandRT = 30,
  bpPARAM = FALSE, LCPUs = 1L)
}
\arguments{
\item{subdir}{directory}

\item{object}{a object of featureChroma}

\item{peakInter}{xx}

\item{expandRT}{rt}

\item{bpPARAM}{xx}

\item{LCPUs}{the number of logical CPU cores on the current host.}
}
\description{
peakReport
}
\examples{
\dontrun{
peakReport(subdir, object, expandRT = 5)
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
