% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{parameters}
\usage{
.onAttach(...)
}
\description{
A dataset containing peak picking parameters.
}
\details{
The contents are as follows:
\itemize{
  \item method.
  \item ppm.
  \item peakwidth.
  \item snthresh.
  \item prefilter.
  \item mzCenterFun.
  \item intergrate.
  \item noise.
  \item adjustTtime
  \item group.bw
  \item group.mzwid
  \item group.minfac
  \item group.minsamp
  \item group.binsize
}
}
