% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipo.param.R
\name{ipo.param}
\alias{ipo.param}
\title{Paramters of peak extraction}
\usage{
ipo.param(resultPP, resultRG, pf.param = NULL)
}
\arguments{
\item{resultPP}{xx}

\item{resultRG}{xx}

\item{pf.param}{xx}
}
\description{
To read the parameter used in the peakTarget function
}
\examples{
\dontrun{
library(IPO)
library(peakFinder)
msdata <- list.files(getwd(),"mzML", full.names=TRUE)
paramsPP <- getDefaultXcmsSetStartingParams()
paramsPP$mzdiff <- -0.001
paramsPP$ppm <- c(5,15)
paramsPP$min_peakwidth <- c(7,14)
paramsPP$max_peakwidth <- c(20,60)
paramsPP$noise <- 100000
paramsPP$snthresh <- c(2,15)
resultPP <- optimizeXcmsSet(msdata[2:3], paramsPP, subdir="test")
paramsRG <- getDefaultRetGroupStartingParams()
paramsRG$gapInit <- 0.2
paramsRG$profStep <- 1
paramsRG$minfrac <- 0.75
resultRG <- optimizeRetGroup(resultPP$best_settings$xset, paramsRG, nSlaves=2)
# ipo----peakFinder
inhouseParam <- ipo.param(resultPP,resultRG,pf.param= NULL)
peak <- peakFinder(platformPara = inhouseParam,bpPARAM = T)
xy <- featureChroma(peak,expandRt = 10,bpPARAM = T)
peak.table(peak,fileName = "peakMatrix")
peakReport(subdir = "chromatography",object = xy, peakInter = NULL,expandRT = 10)
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
