% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakFinder.R
\name{peakFinder}
\alias{peakFinder}
\title{Easy peaks extraction from mass spectroemtry data}
\usage{
peakFinder(files = NULL, platformPara, bpPARAM = FALSE, LCPUs = 1L)
}
\arguments{
\item{files}{if files is NULL, a file selection dialog appears. Alternatively,
a character vectors, containing one of MS data files should be added.}

\item{platformPara}{The parameters of peaks extraction with centWave in XCMS.
A file with list format for in-house parameters.}

\item{bpPARAM}{xx}

\item{LCPUs}{if possible, the number of logical CPUs (if bpPARAM TRUE)}
}
\value{
A peakTarget object. The object includes platformPara information,
a peakTable (the output file format of diffreports in xcms package),
and a XCMSnExp objects. See the details at https://stattarget.github.io
}
\description{
Peaks extraction from high resulution mass spectrometry data,
Optimized parameters are provided for the data produced from UPLC-ORBIRAP,
HPLC-ORBITRAP, HPLC-TOF and UPLC-TOF platform.
}
\examples{
\dontrun{
#read the in-house parameter file
path <- system.file("parameters",package = "peakFinder")
param <- list.files(path,".pk.param.xls",recursive=FALSE,full.names=TRUE)
inhouseParam <- readPKparam(param)
#perform the peakFinder
peakTable <- peakFinder(platformPara = inhouseParam)
#peakTable <- peakFinder(platformPara = inhouseParam,bpPARAM = T)
xy <- featureChroma(peakTable,expandRt = 5,bpPARAM = T)
peakReport(peakTable,fileName = "peakMatrix")
chromatogramReport(subdir = "chromatography",object = xy,expandRT = 5)
}
}
\keyword{peakFinder,}
\keyword{peak_inHouse}
\keyword{peaks,}
