% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatogramReport.R
\name{peakInter}
\alias{peakInter}
\title{peakInter}
\usage{
peakInter(object, divg, topIndex)
}
\arguments{
\item{object}{directory}

\item{divg}{a object of featureChroma}

\item{topIndex}{xx}
}
\description{
peakInter
}
\examples{
\dontrun{
peakReport(subdir, object, expandRT = 5)
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
