% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatogramReport.R
\name{peakReport}
\alias{peakReport}
\title{peakReport}
\usage{
peakReport(object, fileName)
}
\arguments{
\item{object}{a object of peakTarget}

\item{fileName}{file name}
}
\value{
Two tables with .ok.param.tsv and .tsv format
}
\description{
peakReport
}
\examples{
\dontrun{
peak.table(object,fileName)
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
