% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\docType{class}
\name{peakTarget-class}
\alias{peakTarget-class}
\title{An S4 class to represent the parmater for peakTarget}
\value{
A object of peakTarget
}
\description{
An S4 class to represent the parmater for peakTarget
}
\section{Slots}{

\describe{
\item{\code{inputPara}}{The peak picking and anotation parameters.}

\item{\code{filterPeaks}}{The filtered peak IDs as isotope peaks or others.}

\item{\code{peakTable}}{Generate a matrix of integrated peak intensity with rows for every group
and columns for every sample.}

\item{\code{peakChroma}}{a XCMSnExp object contains the results of a G/LC-MS data
preprocessing that comprises chromatographic peak detection, alignment and
correspondence.}
}}

