% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChroma.R
\name{plotChroma}
\alias{plotChroma}
\title{plotChroma}
\usage{
plotChroma(object, num, expandRT = 15, facet_wrap = TRUE,
  leg.pos = "none", SavGolay = FALSE, SavGolay.W = 1/3)
}
\arguments{
\item{object}{a object of featureChroma}

\item{num}{peaks ID}

\item{expandRT}{RT}

\item{facet_wrap}{facet_wrap}

\item{leg.pos}{xx}

\item{SavGolay}{Smooth data with a Savitzky-Golay smoothing filter from signal 0.7-6 package.}

\item{SavGolay.W}{critical frequencies of the Savitzky-Golay smoothing filter filter. W must be a scalar for low-pass and high-pass filters, and W must be a two-element vector c(low, high) specifying the lower and upper bands. For digital filters, W must be between 0 and 1 where 1 is the Nyquist frequency.}
}
\description{
plotChroma
}
\examples{
\dontrun{
plotChroma(xx)
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
