% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPKparam.R
\name{readPKparam}
\alias{readPKparam}
\title{Paramters of peak extraction}
\usage{
readPKparam(x = NULL)
}
\arguments{
\item{x}{a template with '.pk.param.xls' format. if files is NULL, a file selection dialog appears.}
}
\description{
To read the parameter used in the peakTarget function
}
\examples{
\dontrun{
readPKparam()
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
