% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeASChroma.R
\name{writeASChroma}
\alias{writeASChroma}
\title{writeASChroma}
\usage{
writeASChroma(x)
}
\arguments{
\item{x}{a object of featureChroma}
}
\description{
writeASChroma
}
\examples{
\dontrun{
writeASChroma(xx)
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
