% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPKparam.R
\name{writePKparam}
\alias{writePKparam}
\title{Paramters of peak extraction}
\usage{
writePKparam(object, paramfile)
}
\arguments{
\item{object}{a template with '.pk.param.xls' format}

\item{paramfile}{xx}
}
\description{
To read the parameter used in the peakTarget function
}
\examples{
\dontrun{
writePKparam()
}
}
\keyword{peakTarget,}
\keyword{peaks,}
\keyword{readPKparam}
