% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfinder.R
\name{peakfinder}
\alias{peakfinder}
\title{Interactively determine the peak of a densitiy estimation}
\usage{
peakfinder(x)
}
\arguments{
\item{x}{a time series}
}
\description{
A data frame with peak coordinates is shown in the second tab. Press 'Done'
to return the data frame to the R console.
}
\examples{

peakfinder(AirPassengers)
}
